refactor: Rework app sidebar by consolidating navigation components and updating user avatar styling.
