--- components/common/app-sidebar.tsx
@@ -1,308 +1,0 @@
-"use client";
-
-import * as React from "react";
-
-import { NavMain } from "@/components/common/nav-main";
-import { NavProjects } from "@/components/common/nav-projects";
-import {
-  PiArchiveDuotone,
-  PiBookmarkSimpleDuotone,
-  PiChatTeardropDotsDuotone,
-  PiCirclesThreePlusDuotone,
-  PiDiamondsFourDuotone,
-  PiFolderOpenDuotone,
-  PiGearDuotone,
-  PiKanbanDuotone,
-  PiLifebuoyDuotone,
-  PiStackPlusDuotone,
-} from "react-icons/pi";
-
-import {
-  Sidebar,
-  SidebarContent,
-  SidebarGroup,
-  SidebarGroupContent,
-  SidebarHeader,
-  SidebarMenu,
-  SidebarMenuItem,
-  SidebarRail,
-  useSidebar,
-} from "@/components/ui/sidebar";
-import { useGetAllProducts } from "@/hooks/product/useGetAllProducts";
-import { useGetWorkspace } from "@/hooks/workspace/useGetWorkspace";
-import { cn } from "@/lib/utils";
-import { Product } from "@/types/product";
-import {
-  ArrowLeftIcon,
-  BookOpenCheck,
-  Grid2X2,
-  ImagePlus,
-  LayoutGrid,
-  LayoutList,
-  Sheet,
-  Tags,
-} from "lucide-react";
-import Image from "next/image";
-import Link from "next/link";
-import { useParams, usePathname } from "next/navigation";
-import { Button } from "../ui/button";
-
-const data = {
-  navMain: [
-    {
-      title: "Dashboard",
-      url: "/dashboard",
-      icon: PiDiamondsFourDuotone,
-    },
-    {
-      title: "Departments",
-      url: "/departments",
-      icon: PiCirclesThreePlusDuotone,
-    },
-    {
-      title: "Projects",
-      url: "/projects",
-      icon: PiKanbanDuotone,
-    },
-    {
-      title: "Products",
-      url: "/products",
-      icon: PiStackPlusDuotone,
-    },
-    {
-      title: "Source Files",
-      url: "/source-files",
-      icon: PiFolderOpenDuotone,
-    },
-    {
-      title: "Bookmarked Products",
-      url: "/bookmarked-products",
-      icon: PiBookmarkSimpleDuotone,
-    },
-    {
-      title: "Archive",
-      url: "/archive",
-      icon: PiArchiveDuotone,
-    },
-    {
-      title: "Settings",
-      url: "/settings",
-      icon: PiGearDuotone,
-    },
-  ],
-  other: [
-    {
-      name: "Feedback",
-      url: "/feedback",
-      icon: PiChatTeardropDotsDuotone,
-    },
-    {
-      name: "Help Center",
-      url: "/help-center",
-      icon: PiLifebuoyDuotone,
-    },
-  ],
-  product: [
-    {
-      title: "Product information",
-      url: "/product-information",
-      icon: LayoutGrid,
-    },
-    {
-      title: "Compliance information",
-      url: "/compliance-information",
-      icon: BookOpenCheck,
-    },
-    {
-      title: "Label Components",
-      url: "/label-components",
-      icon: LayoutList,
-    },
-    {
-      title: "Symbols & Graphics",
-      url: "/symbols-graphics",
-      icon: ImagePlus,
-    },
-    {
-      title: "Product data",
-      url: "/product-data",
-      icon: Sheet,
-    },
-    {
-      title: "Operational parameters",
-      url: "/operational-parameters",
-      icon: Grid2X2,
-    },
-    {
-      title: "Label tags",
-      url: "/label-tags",
-      icon: Tags,
-    },
-  ],
-};
-
-const team = {
-  name: "Uprevit",
-  company: "Uprevit Inc",
-  logoSrc: "/avatars/workspace-logo.png",
-};
-
-export function AppSidebar({ ...props }: React.ComponentProps<typeof Sidebar>) {
-  const params = useParams();
-  const pathname = usePathname();
-  const { open } = useSidebar();
-  const { data: workspaceData, isLoading } = useGetWorkspace();
-  const workspace = workspaceData?.workspace;
-
-  const productId =
-    typeof params.productId === "string"
-      ? params.productId
-      : Array.isArray(params.productId)
-      ? params.productId[0]
-      : undefined;
-
-  const isProductPage = Boolean(productId);
-
-  const { data: productsData } = useGetAllProducts();
-  const product = productId
-    ? productsData?.result.products?.find((p: Product) => p._id === productId)
-    : undefined;
-
-  return (
-    <Sidebar
-      collapsible="icon"
-      className={cn("overflow-hidden [&>[data-sidebar=sidebar]]:flex-row")}
-      {...props}
-    >
-      {/* This is the first sidebar */}
-      {/* We disable collapsible and adjust width to icon. */}
-      {/* This will make the sidebar appear as icons. */}
-      <Sidebar
-        collapsible={isProductPage ? "none" : "icon"}
-        className={cn(
-          isProductPage && "!w-[calc(var(--sidebar-width-icon)_+_1px)] border-r"
-        )}
-      >
-        <SidebarHeader>
-          <Link
-            href="/"
-            className="flex items-center gap-4 p-1 rounded  data-[state=open]:bg-sidebar-accent data-[state=open]:text-sidebar-accent-foreground"
-          >
-            <div className="relative flex aspect-square mb-1 size-6 items-center justify-center">
-              <Image
-                src="/log-no-bg-white.svg"
-                alt="Uprevit logo"
-                fill
-                className=""
-              />
-            </div>
-            <div className="grid flex-1 text-left text-sm leading-tight">
-              <span className="truncate text-lg text-background font-black ">
-                UPREVIT
-              </span>
-            </div>
-          </Link>
-        </SidebarHeader>
-        <SidebarContent>
-          <NavMain
-            items={data.navMain}
-            open={open}
-            isProductPage={isProductPage}
-          />
-          <NavProjects
-            other={data.other}
-            open={open}
-            isProductPage={isProductPage}
-          />
-        </SidebarContent>
-
-        {/* Workspace */}
-        <SidebarGroup>
-          <SidebarMenu>
-            <SidebarMenuItem className="px-1 ">
-              <Link href={"/settings"}>
-                <button
-                  className={cn(
-                    "cursor-pointer hover:text-sidebar-primary-foreground rounded-lg py-1 flex items-center gap-2 w-full "
-                  )}
-                >
-                  <div className="relative flex size-8 min-w-8 items-center justify-center overflow-hidden rounded-md bg-muted">
-                    <Image
-                      src={team.logoSrc}
-                      alt="Workspace logo"
-                      fill
-                      className="object-cover p-1"
-                    />
-                  </div>
-                  <div className="grid flex-1 text-left leading-tight">
-                    <span className="truncate font-medium">
-                      {isLoading ? "Loading..." : workspace?.workspaceName}
-                    </span>
-                    <span className="truncate text-xs text-muted-foreground">
-                      {isLoading ? "Loading..." : workspace?.companyName}
-                    </span>
-                  </div>
-                </button>
-              </Link>
-            </SidebarMenuItem>
-          </SidebarMenu>
-        </SidebarGroup>
-        <SidebarRail />
-      </Sidebar>
-
-      {/* This is the second sidebar */}
-      {/* We disable collapsible and let it fill remaining space */}
-      {isProductPage && (
-        <Sidebar
-          collapsible="none"
-          className={cn("hidden flex-1 md:flex bg-background")}
-        >
-          <SidebarHeader className="gap-4 border-b p-3">
-            <div className="flex flex-col w-full items-start">
-              <div className="text-sm font-semibold text-foreground">
-                {product?.product_name || "Product"}
-              </div>
-            </div>
-            <div className="flex gap-2 w-full justify-between">
-              <Link href="/products">
-                <Button size="sm" variant="outline">
-                  <ArrowLeftIcon className="size-4" />
-                  Back
-                </Button>
-              </Link>
-            </div>
-          </SidebarHeader>
-          <SidebarContent className="px-0">
-            <SidebarGroup className="px-0">
-              <SidebarGroupContent>
-                <SidebarMenu>
-                  <SidebarMenuItem className="py-2 flex flex-col gap-4">
-                    {data.product &&
-                      data.product.map((item) => (
-                        <Link
-                          key={item.title}
-                          href={`/products/${productId}/${item.url}`}
-                        >
-                          <div
-                            className={cn(
-                              "border-y border-input flex flex-row items-center gap-2 py-5 px-3",
-                              pathname.includes(item.url)
-                                ? "text-primary bg-background"
-                                : ""
-                            )}
-                          >
-                            {item.icon && <item.icon className="size-4" />}
-                            <span>{item.title}</span>
-                          </div>
-                        </Link>
-                      ))}
-                  </SidebarMenuItem>
-                </SidebarMenu>
-              </SidebarGroupContent>
-            </SidebarGroup>
-          </SidebarContent>
-        </Sidebar>
-      )}
-    </Sidebar>
-  );
-}
